cluster/applications/memory-bank/pyproject.toml
[project]
name = "memory-bank"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "uvicorn==0.23.1",
    "asyncio==3.4.3",
    "aiohttp==3.8.4",
    "fastapi==0.115.2",
    "openai==1.79.0",
    "tiktoken==0.7.0",
    "qdrant-client==1.10.1",
    "fastembed==0.6.1",
    "pydantic==2.7.2",
    "pydantic-settings==2.7.1",
    "opentelemetry-api==1.24.0",
    "opentelemetry-sdk==1.24.0",
    "opentelemetry-exporter-otlp==1.24.0",
    "opentelemetry-instrumentation-httpx==0.45b0",
    "opentelemetry-instrumentation-fastapi==0.45b0",
    "opentelemetry-exporter-prometheus==0.45b0",
    "prometheus-client==0.17.0",
    "python-dotenv==1.0.0",
    "python-json-logger==2.0.7",
    "fastapi-mcp>=0.3.3",
]
